move recording apis to their own handler
