+++ viseron/components/webserver/api/v1/__init__.py
@@ -3,9 +3,11 @@
 from viseron.components.webserver.api.v1.camera import CameraAPIHandler
 from viseron.components.webserver.api.v1.cameras import CamerasAPIHandler
 from viseron.components.webserver.api.v1.config import ConfigAPIHandler
+from viseron.components.webserver.api.v1.recordings import RecordingsAPIHandler
 
 __all__ = (
     "CameraAPIHandler",
     "CamerasAPIHandler",
     "ConfigAPIHandler",
+    "RecordingsAPIHandler",
 )
